+++ GameShared/packet.hpp
@@ -28,7 +28,7 @@ protected:
 
 protected: // Packet Data
     PacketType type;
-    unsigned long int connection_id;
+    unsigned int connection_id;
     unsigned int sequence;
     unsigned int ack;
     unsigned int ack_bitfield;
@@ -42,12 +42,12 @@ public:
     virtual void Decode(char* buffer);
 
     PacketType GetType() { return this->type; }
-    unsigned long int GetConnectionId() { return this->connection_id; }
+    unsigned int GetConnectionId() { return this->connection_id; }
     unsigned int GetSequence() { return this->sequence; }
     unsigned int GetAck() { return this->ack; }
     unsigned int GetAckBitfield() { return this->ack_bitfield; }
 
-    void SetConnectionId(unsigned long int connection_id) { this->connection_id = connection_id; }
+    void SetConnectionId(unsigned int connection_id) { this->connection_id = connection_id; }
     void SetSequence(unsigned int sequence) { this->sequence = sequence; }
     void SetAck(unsigned int ack) { this->ack = ack; }
     void SetAckBitfield(unsigned int ack_bitfield) { this->ack_bitfield = ack_bitfield; }
